requirements.txt
pandas>=2.0
numpy>=1.24
pydantic>=2.0
scipy>=1.10
scikit-learn>=1.3
lightgbm>=4.0
httpx>=0.25
python-dotenv>=1.0
PyYAML>=6.0
requests>=2.28
requests-cache>=1.2
retry-requests>=2.0
openmeteo-requests>=1.2
# ECMWF GRIB(data_pipeline/ecmwf_forecast_data.py) — 선택 설치, Windows 는 eccodes 까다로울 수 있음
# ecmwf-opendata>=0.3.0
# ecmwf-api-client>=1.6.0
# xarray>=2024.1.0
# cfgrib>=0.9.10
